fix: fixed warning when starting building
